+++ e-commerce/src/main/java/com/example/e_commerce/repository/OrderItemRepository.java
@@ -1,10 +1,12 @@
 package com.example.e_commerce.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.e_commerce.entity.OrderItem;
-import com.example.e_commerce.entity.Order;
 import java.util.List;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.e_commerce.entity.Order;
+import com.example.e_commerce.entity.OrderItem;
+
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
     List<OrderItem> findByOrder(Order order);
 }
