+++ e-commerce/src/main/java/com/example/e_commerce/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.example.e_commerce.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.e_commerce.entity.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 }
